requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.1.0
bytecode==0.13.0
cattrs==22.2.0
celery==5.2.7
celery-redbeat==2.0.0
coverage==7.1.0
ddsketch==2.0.4
ddtrace==1.5.1
Deprecated==1.2.13
Django==4.2.7
django-celery-beat==2.5.0
django-celery-results==2.4.0
django-environ==0.9.0
django-redis==5.2.0
djangorestframework==3.14.0
drf-spectacular==0.24.2
envier==0.4.0
exceptiongroup==1.0.0rc9
gevent==22.10.2
gunicorn==20.1.0
ipaddress==1.0.23
jsonschema==4.16.0
packaging==21.3
protobuf==4.21.7
psycopg2-binary==2.9.5
pyparsing==3.0.9
pyrsistent==0.18.1
pytz==2022.4
pyyaml==6.0
redis==4.3.4
six==1.16.0
sqlparse==0.4.4
tenacity==8.1.0
typing_extensions==4.4.0
wrapt==1.14.1
uritemplate==4.1.1
xmltodict==0.13.0
drf_spectacular==0.24.2
PyJWT~=2.6.0
cryptography~=41.0.4
python-dotenv~=0.21.0
django-cors-headers==3.13.0
pysnowflake~=0.1.3
pytest~=7.2.0
snowflake-connector-python~=3.1.0
msal~=1.23.0
pandas~=1.5.0
python-dateutil~=2.8.2